Fix treatment of nil pointer values.
This addresses an issue where pointer fields were always being
initialized to a zero value.

Now if a struct has a field that is missing from the query results, that
field will maintain its original nil value if it is a pointer. Likewise,
query results that are NULL will do the same if the PtrNil client option
is set.